+++ R/Read_Results.R
@@ -9,8 +9,6 @@
 #' @importFrom rvest html_text
 #' @importFrom pdftools pdf_text
 #'
-#'
-#'
 #' @param file a .pdf or .html file (could be a url) where containing swimming results.  Must be formatted in a "normal" fashion - see vignette
 #' @param node a CSS node where html results are stored.  Required for html results.
 #' @return returns a list of strings containing the information from \code{x}.  Should then be parsed with \code{Swim_Parse}
@@ -42,8 +40,14 @@ Read_Results <- function(file, node = NULL) {
     stop("Please supply a valid .html or .pdf document")
   }
 
+  # extracts lines as list
   as_lines <- str_extract_all(results, "\n.*")
   as_lines_list_2 <- unlist(as_lines, recursive = FALSE)
 
   return(as_lines_list_2)
 }
+
+
+#' @rdname Read_Results
+#' @export
+read_results <- Read_Results